Speed up the tests
Instead of using an arbitrary time to wait, instead use a full rendering
cycle (3 animation frames) when we want to wait for a render to have
occured. In the future we can make this more granular like "wait for the
render step to complete" or "wait for the commit step to complete".
Might even be able to tie it directly into the scheduler, but for now
this is an improvement.