Use MutationObserver to teardown virtual components
Directives do not have a lifecycle API. So to know when they are
removed, and call the appropriate useEffect teardown functions we now
use MutationObserver.

We do a little dance that goes like this:

* MutationObserver for when the directive node is removed from its
parent DocumentFragment (internal to lit-html).
* MutationObserver for when the directive is inserted into shadow DOM
(or whatever else I suppose).
* MutationObserver for when removed.

The last is when teardown is called. Fixes #69